requirements.txt
Flask
python-dotenv
psycopg2-binary
Flask-JWT-Extended
